chore: use tsdown for bundling
