+++ vite.config-common.mts
@@ -1,7 +1,12 @@
-import { defineConfig, type Plugin, withFilter, type PluginOption } from "vite"
+import { defineConfig, type PluginOption } from "vite"
 import pluginReact from "@vitejs/plugin-react"
-import pluginSwc from "@rollup/plugin-swc"
 import { sentryVitePlugin } from "@sentry/vite-plugin"
+import {
+    BUILD_TARGET,
+    pluginOptimizeReactAriaLocales,
+    pluginSwcDecorators,
+    scssPreprocessorOptions,
+} from "./rolldown.config-common.mts"
 import * as clientSettings from "./settings/clientSettings.js"
 import {
     VITE_ASSET_SITE_ENTRY,
@@ -9,33 +14,10 @@ import {
     ViteEntryPoint,
 } from "./site/viteConstants.js"
 
-// see docs/browser-support.md
-export const BUILD_TARGET = ["chrome106", "firefox110", "safari16.0"]
+export { BUILD_TARGET }
 
 export const commonPlugins = (): PluginOption[] => [
-    withFilter(
-        // Use swc to transform decorators, since rolldown/oxc doesn't support modern decorators yet. We could remove this once they do - see https://github.com/oxc-project/oxc/issues/9170.
-        pluginSwc({
-            swc: {
-                jsc: {
-                    parser: {
-                        syntax: "typescript",
-                        decorators: true,
-                    },
-                    transform: {
-                        decoratorVersion: "2023-11",
-                        useDefineForClassFields: true,
-                    },
-
-                    // This setting we need to override from @rollup/plugin-swc's default, otherwise it will not put optional properties on classes (e.g. `class A { optionalProp?: string }`), thereby breaking mobx decorators
-                    loose: false,
-                    target: "esnext",
-                },
-            },
-        }),
-        // Only run this transform if the file contains a decorator.
-        { transform: { code: /[^"]@/, id: /.*\.(ts|tsx)$/ } }
-    ),
+    pluginSwcDecorators(),
     pluginReact(),
     pluginOptimizeReactAriaLocales({
         locales: ["en-US"],
@@ -55,25 +37,7 @@ export const defineViteConfigForEntrypoint = (entrypoint: ViteEntryPoint) => {
         css: {
             devSourcemap: true,
             preprocessorOptions: {
-                scss: {
-                    // Prevent reintroducing deprecated features.
-                    fatalDeprecations: [
-                        "color-functions",
-                        "global-builtin",
-                        "mixed-decls",
-                        "slash-div",
-                    ],
-                    quietDeps: true,
-                    silenceDeprecations: [
-                        // We don't want to deal with the import warnings for now.
-                        // https://sass-lang.com/documentation/breaking-changes/import/
-                        //
-                        // Some of them come from dependencies. For example,
-                        // they should be fixed in the upcoming Bootstrap 6.
-                        // https://github.com/twbs/bootstrap/issues/29853
-                        "import",
-                    ],
-                },
+                scss: scssPreprocessorOptions,
             },
         },
         define: Object.fromEntries(
@@ -145,81 +109,4 @@ export const defineViteConfigForEntrypoint = (entrypoint: ViteEntryPoint) => {
             port: vitePort,
         },
     })
-}
-
-// This plugin removes locale imports from react-aria packages.
-// It is a copy of https://github.com/adobe/react-spectrum/tree/0a84129f133bc549df31ad4be17a2fe6a9bceed4/packages/dev/optimize-locales-plugin (available as @react-aria/optimize-locales-plugin on npm),
-// but fixed to work correctly, and optimized (using Rolldown-native filtering).
-const pluginOptimizeReactAriaLocales = ({
-    locales,
-}: {
-    locales: readonly string[]
-}): Plugin => {
-    const emptyLocaleModuleId = "\0owid-empty-react-aria-locale"
-    const emptyLocaleModule = `export default undefined;`
-    const reactAriaPackagePathRegex =
-        /[/\\](?:@?react-stately|@?react-aria|@?react-spectrum|@?react-aria-components)[/\\]/
-    const localeImportSpecifierRegex =
-        /(?:^|[/\\])([a-z]{2}-[A-Z]{2})(?:\.(?:[cm]?js|json))?(?:[?#].*)?$/
-
-    const getLocaleFromFilename = (specifier: string): string | undefined => {
-        return specifier.match(localeImportSpecifierRegex)?.[1]
-    }
-
-    const getIntlLocale = (localeName: string): Intl.Locale | undefined => {
-        try {
-            return new Intl.Locale(localeName)
-        } catch {
-            return undefined
-        }
-    }
-
-    const localeMatches = (
-        localeToMatch: Intl.Locale,
-        includedLocale: Intl.Locale
-    ): boolean =>
-        localeToMatch.language === includedLocale.language &&
-        (!includedLocale.region ||
-            localeToMatch.region === includedLocale.region)
-
-    const includedLocales = locales.map((locale) => new Intl.Locale(locale))
-
-    return {
-        name: "owid-optimize-react-aria-locales",
-        enforce: "pre",
-        resolveId: {
-            filter: { id: localeImportSpecifierRegex },
-            handler(source, importer, options) {
-                if (
-                    !importer ||
-                    options.ssr ||
-                    !reactAriaPackagePathRegex.test(importer)
-                )
-                    return null
-
-                const localeName = getLocaleFromFilename(source)
-                if (!localeName) return null
-
-                const locale = getIntlLocale(localeName)
-                if (!locale) return null
-
-                if (
-                    includedLocales.some((includedLocale) =>
-                        localeMatches(locale, includedLocale)
-                    )
-                )
-                    return null
-
-                return emptyLocaleModuleId
-            },
-        },
-        load: {
-            filter: { id: new RegExp(`^${emptyLocaleModuleId}$`) },
-            handler(id) {
-                if (id === emptyLocaleModuleId) return emptyLocaleModule
-
-                return null
-            },
-        },
-    }
 }
